requirements.txt
Django==4.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
psycopg2-binary==2.9.*
cloudinary==1.36.*
django-cloudinary-storage==0.3.*
africastalking==1.2.*
resend==0.8.*
python-decouple==3.8
gunicorn==21.2.*
whitenoise==6.6.*
Pillow==10.2.*
dj-database-url==2.1.*
sentry-sdk==1.40.*
celery==5.3.*
redis==5.0.*
django-redis==5.4.*
